requirements.txt
absl-py
altair
astunparse
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
commonmark
decorator
entrypoints
flatbuffers
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
Jinja2
jsonschema
keras
Keras-Preprocessing
libclang
Markdown
MarkupSafe
numpy
oauthlib
opt-einsum
packaging
pandas
Pillow
protobuf
pyarrow
pyasn1
pyasn1-modules
pydeck
Pygments
Pympler
pyparsing
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
requests
requests-oauthlib
rich
rsa
semver
six
smmap
streamlit
streamlit-lottie
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
toml
toolz
tornado
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
Werkzeug
wrapt
zipp
